membership-products: deduplicate connection notice across resolvers
Address reviewer feedback: extract handleResolverError into utils.js so
all resolvers (getProducts, getSubscriberCounts, getNewsletterCategories,
getPostEmailSentState) share the same rest_unauthorized guard with a
stable notice id for deduplication. Remove unused connect_url addition
from proxy_request_to_wpcom since the JS side builds the URL client-side.
Update changelog to clarify scope covers all Jetpack sites, not just Atomic.
