--- projects/packages/connection/changelog/dotsite-127-rest-unauthorized-connect-url
@@ -1,4 +1,0 @@
-Significance: patch
-Type: changed
-
-Include `connect_url` in the WP_Error data returned by `proxy_request_to_wpcom` when the current user is not connected to WordPress.com, so API consumers can use it to build a connection flow.
