+++ projects/plugins/jetpack/extensions/store/membership-products/resolvers.js
@@ -1,9 +1,6 @@
-import { getUserConnectionUrl } from '@automattic/jetpack-connection';
 import { isSimpleSite } from '@automattic/jetpack-script-data';
 import apiFetch from '@wordpress/api-fetch';
 import { store as editorStore } from '@wordpress/editor';
-import { __ } from '@wordpress/i18n';
-import { store as noticesStore } from '@wordpress/notices';
 import { addQueryArgs, getQueryArg } from '@wordpress/url';
 import { PRODUCT_TYPE_PAYMENT_PLAN } from '../../shared/components/product-management-controls/constants';
 import { getMessageByProductType } from '../../shared/components/product-management-controls/utils';
@@ -22,7 +19,7 @@ import {
 	setPostEmailSentState,
 } from './actions';
 import { API_STATE_CONNECTED, API_STATE_NOTCONNECTED } from './constants';
-import { onError } from './utils';
+import { handleResolverError } from './utils';
 
 const EXECUTION_KEY = 'membership-products-resolver-getProducts';
 const SUBSCRIBER_COUNT_EXECUTION_KEY = 'membership-products-resolver-getSubscriberCounts';
@@ -257,30 +254,7 @@ export const getProducts =
 		} catch ( error ) {
 			dispatch( setConnectUrl( null ) );
 			dispatch( setApiState( API_STATE_NOTCONNECTED ) );
-
-			if ( error?.code === 'rest_unauthorized' && ! isSimpleSite() ) {
-				const connectUrl = getUserConnectionUrl( { from: 'editor' } );
-				registry
-					.dispatch( noticesStore )
-					.createNotice(
-						'warning',
-						__(
-							'To use publishing features like subscriptions and paid memberships, connect your WordPress.com account.',
-							'jetpack'
-						),
-						{
-							id: 'jetpack-memberships-user-connection-required',
-							actions: [
-								{
-									label: __( 'Connect account', 'jetpack' ),
-									url: connectUrl,
-								},
-							],
-						}
-					);
-			} else {
-				onError( error.message, registry );
-			}
+			handleResolverError( error, registry );
 		} finally {
 			executionLock.release( lock );
 		}
@@ -304,7 +278,7 @@ export const getSubscriberCounts =
 			);
 		} catch ( error ) {
 			dispatch( setApiState( API_STATE_NOTCONNECTED ) );
-			onError( error.message, registry );
+			handleResolverError( error, registry );
 		} finally {
 			executionLock.release( lock );
 		}
@@ -347,7 +321,7 @@ export const getNewsletterCategories =
 			);
 		} catch ( error ) {
 			dispatch( setApiState( API_STATE_NOTCONNECTED ) );
-			onError( error.message, registry );
+			handleResolverError( error, registry );
 		} finally {
 			executionLock.release( lock );
 		}
@@ -372,7 +346,7 @@ export const getPostEmailSentState =
 				} )
 			);
 		} catch ( error ) {
-			onError( error.message, registry );
+			handleResolverError( error, registry );
 		} finally {
 			executionLock.release( lock );
 		}
